+++ src/qcir/oracle/deancilla.cpp
